商品发布

+++ toucan-shopping-modules/product/toucan-shopping-product-entity-vo/src/main/java/com/toucan/shopping/modules/product/entity/ShopProduct.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @Data
 public class ShopProduct {
     private Long id; //主键
-    private Integer categoryId; //所属类别
+    private Long categoryId; //所属类别
     private Long shopCategoryId; //店铺内的所属分类
     private String uuid; //SPU UUID
     private String name; //商品名称
